Optimized anonymous mode
If anonymous mode is on, then the teacher should not be able to reveal
the identities, also not afterwards.
So I made a change to the sessionattempts table that the user ist not
linked and his identity stays anonymous. This is implemented for both
modes: single and group mode.

To be unable to get the identities revealed, the grade function should
also not be selectable unless the trainer could match answers with
users in the gradebook.